Add problem 51 solution
This is a little verbose. We construct a sieve with which to iterate
over primes and determine the digits which can be replaced within the
prime. These are then added to a hashmap which is queried for the first
8 digit family.

We do a small check to iterate over all values in a given sequence from
[10^t, 10^(t-1)] since the family with the smallest prime might not
necessarily be found first.