Add problem 93 solution
This took a little longer than it should have. It took a little too long
to realise that the values being produced were not using the standard
integer floored division as is common in programming languages, but
instead we could get floats back.

A standard brute-force approach otherwise, with enough macros to make it
fairly readable.